Added a CLI to make generated exes more useful.
Created a command line interface (CLI) that allows no-compile means of running late entries.

The "burning platform" for this is an attempt to help Magenta, who doesn't know C++ well and can't compile this, but just wants to use the executable to test his Python entry (and train his bot).

The pool class was changed to store a private vector of function entries for use with each applicant, rather than hard coding the factory in the newPlayer method.  This allows registration to be controlled a bit more dynamically.  The new way to register is to add a REGISTER_ENTRY(type); macro call in main.  Note that although this entry process should be easier, the main reason for it is to support dynamic registration flexibility for the CLI.

This change allows the CLI options to create a "late entry", and set its NL style if required.  Some help text was added as well for this.

In addition, the command line allows a user to completely bypass registering all tournament entries, instead adding HumanPlayer.  Using this option, a person can debug his own ChildProcessPlayer candidate.